requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
python-slugify==8.0.4
SQLAlchemy==2.0.44
text-unidecode==1.3
typing_extensions==4.15.0
Werkzeug==3.1.4
flask-mail
psycopg2-binary
python-dotenv